Restructure popup tests: co-locate with source, flatten describes, use "should" naming
